+++ src/vars/commandData.ts
@@ -1145,7 +1145,7 @@ const osucmds: bottypes.commandInfo[] = [
 
             }
         ],
-        aliases: [],
+        aliases: ['pins'],
         args: scoreListCommandOptions
     },
     {
@@ -1408,7 +1408,17 @@ const osucmds: bottypes.commandInfo[] = [
                 description: 'Shows your recent scores with the map name/difficulty/artist/creator matching "shinbatsu", excluding fails'
             }
         ],
-        aliases: ['rs', 'r', 'rt', 'rf', 'rm', 'rctb', 'rl', 'rlt', 'rlf', 'rlm', 'rlctb', 'rsbest', 'recentbest', 'rb'],
+        aliases: [
+            'recentscore', 'rs', 'r',
+            'recenttaiko', 'rt',
+            'recentfruits', 'rf', 'rctb',
+            'recentmania', 'rm',
+            'rb', 'recentbest', 'rsbest',
+            'rslist', 'recentlist', 'rl',
+            'recentlisttaiko', 'rlt',
+            'recentlistfruits', 'rlf', 'rlctb', 'rlc',
+            'recentlistmania', 'rlm',
+        ],
         args: scoreListCommandOptions.slice(0, 9).concat(
             scoreListCommandOptions.slice(11))
             .concat([